requirements.txt
torch==2.9.1
vision==0.24.1
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==12.0.0
pytest==7.4.3
numpy==2.3.5
matplotlib==3.10.7